Make a small change on the both models: remove unique constraint on names
Channels can has names on multiple languages, or aliases. A unique
identifier is sufficient.

Category has a big problem if name has a unique constraint. As
Category is tree-like structure, a Category can be a instance of
different parents. Like this:

Books / Computers,
Computers / Notebooks,

We have a category(name=Computers, parent=Books) and
category(name=Category, parent=None), that categories are different!
But as was defined as constraint UNIQUE on name before this is not
possible. So I'm removing that.