Change behavior of core.views.CategoryDetailView
Instead return multiple instances of different pairs of
(Parent, Category) as a concated result of their subcategories
and parents.

Example: (None, Computers) and (Book, Computers)

Are in different Category depths, so has parents and subcategories
differents as well. But using core.views.CategoryDetailView._concat
The results of underlying parents and subcategories are concatened