Apply recursive serializer on views: ChannelDetail and CategoryDetail
Implementation is based on the serializers.CategoryRootSerializer
whose make a self-referentiation serialization through all the
subcategories.

This affect the CategoryDetailSerializer on the
subcategories fields. Now the API returns a nested JSON for the given
categories. The parents fields was not modified.

After that the parent_name field was deleted because is useless.