Improvement on _read_csv of importcategories command
On previous version that command only will work if the csv file
has "Category" column as its first column, as is the categories.csv
stored on that repository.

However, on test description was said that:

"... Each channel sends us a CSV file where one
of the columns (`Category`) is .. "

TAG: [WHERE ONE OF THE COLUMNS]

So the file may have multiple columns.

In this implementation the routine make a search to lookup
in which index the Category column is and so then return each cell
of this column as before.